backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
librosa==0.10.1
SpeechRecognition==3.10.1
numpy==1.26.4
soundfile==0.12.1
werkzeug==3.0.1
langdetect==1.0.9
